requirements.txt
pytest
allure-pytest
playwright